Make const reassignments only a warning (#6020)
