+++ test/function/samples/warning-const-reassign/main.js
